% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guarded.R
\name{\%guarded\%}
\alias{\%guarded\%}
\title{Defines a function with guards}
\usage{
lhs \%guarded\% e
}
\value{
A compiled, guarded function
}
\description{
is an  operator which is used to define functions with guard statements
}
\details{
Guard is an boolean expression in parenthesis, followed by the expression to execute
if the guard expression evals to true.
(y!=0)
{
  x/y
}

Both, the guard and function expression can be any valid R code, as long as the
syntax remains the same.

Internally, guard simply compiles the function body to a  if (...) else if (...) else
statement.
}
\examples{
fib <- .\%guarded\% function(x){
	(x<=1)
	{
		x
	}
	(otherwise)
	{
		Recall(x-1) + Recall(x-2)
	}
}
}
