% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guarded.R
\docType{data}
\name{otherwise}
\alias{otherwise}
\title{Otherwise guard}
\format{An object of class \code{logical} of length 1.}
\usage{
otherwise
}
\description{
Use as the last guard statement (if required)
}
\examples{
(otherwise)
{
   NA
}
}
\keyword{datasets}
